Refactor estimated position layer to not display when vessels groups are only displayed
