+++ frontend/src/features/Vessel/layers/VesselsEstimatedPositionLayer/utils.ts
@@ -11,7 +11,6 @@ import { transform } from 'ol/proj'
 export const getEstimatedPositionFeatures = (
   vessel: Vessel.VesselLastPosition,
   options: {
-    isHidden: boolean
     isLight: boolean
     vesselIsHidden: Date
     vesselIsOpacityReduced: Date
@@ -45,7 +44,6 @@ export const getEstimatedPositionFeatures = (
   const lineFeature = new Feature({
     color: lineColor,
     geometry: new LineString([currentCoordinates, estimatedCoordinates]),
-    isHidden: options.isHidden,
     opacity
   }) as Vessel.VesselEstimatedPositionFeature
   lineFeature.estimatedPosition = estimatedPositionData
@@ -55,7 +53,6 @@ export const getEstimatedPositionFeatures = (
     color: vesselColor,
     geometry: new Point(estimatedCoordinates),
     isCircle: true,
-    isHidden: options.isHidden,
     opacity
   }) as Vessel.VesselEstimatedPositionFeature
   circleFeature.estimatedPosition = estimatedPositionData
