+++ frontend/src/features/Vessel/layers/style.ts
@@ -1,15 +1,6 @@
 import { featureHas, featurePropertyIsNotEmpty, stateIs } from '@features/Map/layers/styles/utils/webgl'
-import {
-  VESSEL_ALERT_AND_BEACON_MALFUNCTION,
-  VESSEL_ALERT_STYLE,
-  VESSEL_BEACON_MALFUNCTION_STYLE,
-  VESSEL_INFRACTION_SUSPICION_STYLE,
-  VESSEL_SELECTOR_STYLE,
-  VesselFeature
-} from '@features/Vessel/types/vessel'
+import { VESSEL_SELECTOR_STYLE, VesselFeature } from '@features/Vessel/types/vessel'
 import { Icon, Style } from 'ol/style'
-import Circle from 'ol/style/Circle'
-import Stroke from 'ol/style/Stroke'
 
 import { theme } from '../../../ui/theme'
 import { booleanToInt } from '../../../utils'
@@ -133,73 +124,4 @@ export const getSelectedVesselStyle =
 
 export function degreesToRadian(course) {
   return (course * Math.PI) / 180
-}
-
-const vesselAlertBigCircleStyle = new Style({
-  image: new Icon({
-    src: 'map-icons/Double-cercle-alertes.png'
-  }),
-  zIndex: VESSEL_ALERT_STYLE
-})
-
-export const getVesselAlertStyle = (resolution: number) => {
-  const styles = [vesselAlertBigCircleStyle]
-
-  const scale = Math.min(1, 0.3 + Math.sqrt(200 / resolution))
-  styles[0]?.getImage()?.setScale(scale)
-
-  return styles
-}
-
-const vesselBeaconMalfunctionBigCircleStyle = new Style({
-  image: new Icon({
-    src: 'map-icons/Double-cercle_avaries.png'
-  }),
-  zIndex: VESSEL_BEACON_MALFUNCTION_STYLE
-})
-
-export const getVesselBeaconMalfunctionStyle = (resolution: number) => {
-  const styles = [vesselBeaconMalfunctionBigCircleStyle]
-
-  const scale = Math.min(1, 0.3 + Math.sqrt(200 / resolution))
-  styles[0]?.getImage()?.setScale(scale)
-
-  return styles
-}
-
-const vesselAlertAndBeaconMalfunctionBigCircleStyle = new Style({
-  image: new Icon({
-    src: 'map-icons/Triple-cercle_alerte_et_avarie.png'
-  }),
-  zIndex: VESSEL_ALERT_AND_BEACON_MALFUNCTION
-})
-
-export const getVesselAlertAndBeaconMalfunctionStyle = (resolution: number) => {
-  const styles = [vesselAlertAndBeaconMalfunctionBigCircleStyle]
-
-  const scale = Math.min(1, 0.3 + Math.sqrt(200 / resolution))
-  styles[0]?.getImage()?.setScale(scale)
-
-  return styles
-}
-
-const vesselInfractionSuspicionCircleStyle = new Style({
-  image: new Circle({
-    fill: undefined,
-    radius: 19,
-    stroke: new Stroke({
-      color: theme.color.maximumRed,
-      width: 2
-    })
-  }),
-  zIndex: VESSEL_INFRACTION_SUSPICION_STYLE
-})
-
-export const getVesselInfractionSuspicionStyle = (resolution: number) => {
-  const styles = [vesselInfractionSuspicionCircleStyle]
-
-  const scale = Math.min(1, 0.3 + Math.sqrt(200 / resolution))
-  styles[0]?.getImage()?.setScale(scale)
-
-  return styles
 }
