+++ frontend/src/features/Vessel/utils.ts
@@ -22,6 +22,8 @@ export function buildFeature(
     dateTime: vessel.dateTime,
     detectabilityRiskFactor: vessel.detectabilityRiskFactor,
     emissionPeriod: vessel.emissionPeriod,
+    estimatedCurrentLatitude: vessel.estimatedCurrentLatitude,
+    estimatedCurrentLongitude: vessel.estimatedCurrentLongitude,
     externalReferenceNumber: vessel.externalReferenceNumber,
     flagState: vessel.flagState,
     geometry: new Point(vessel.coordinates),
@@ -42,7 +44,9 @@ export function buildFeature(
     lastControlDateTimeTimestamp: vessel.lastControlDateTimeTimestamp,
     lastLogbookMessageDateTime: vessel.lastLogbookMessageDateTime,
     lastPositionSentAt: vessel.lastPositionSentAt,
+    latitude: vessel.latitude,
     length: vessel.length,
+    longitude: vessel.longitude,
     mmsi: vessel.mmsi,
     numberOfGroupsHidden: vesselGroupDisplayed.numberOfGroupsHidden,
     probabilityRiskFactor: vessel.probabilityRiskFactor,
